Move commands to external file
